Pass integer to get_time

+++ src/accessom_coupler/ocean_solo.F90
@@ -295,7 +295,7 @@ program main
   call get_time(Time_start,ss)
   ! The last sfix time has to be determined from absolute model time, to ensure reproducibility across
   ! restarts
-  Time_last_sfix = set_time(seconds=int(ss/(sfix_hours*SECONDS_PER_HOUR))*sfix_hours*SECONDS_PER_HOUR)
+  Time_last_sfix = set_time(seconds=int(int(ss/(sfix_hours*SECONDS_PER_HOUR))*sfix_hours*SECONDS_PER_HOUR))
   Time_sfix = set_time(seconds=int(sfix_hours*SECONDS_PER_HOUR))
 #endif
